@@ -15,9 +15,9 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
-        <spring-boot.version>2.0.4.RELEASE</spring-boot.version>
-        <spring.cloud.k8s.version>0.3.0.RELEASE</spring.cloud.k8s.version>
+        <java.version>17</java.version>
+        <spring-boot.version>3.2.3</spring-boot.version>
+        <spring.cloud.k8s.version>3.1.0</spring.cloud.k8s.version>
     </properties>
 
     <dependencyManagement>
@@ -42,8 +42,8 @@
             <artifactId>spring-boot-starter-actuator</artifactId>
         </dependency>
         <!--<dependency>-->
-            <!--<groupId>org.springframework.boot</groupId>-->
-            <!--<artifactId>spring-boot-devtools</artifactId>-->
+        <!--<groupId>org.springframework.boot</groupId>-->
+        <!--<artifactId>spring-boot-devtools</artifactId>-->
         <!--</dependency>-->
         <dependency>
             <groupId>de.codecentric</groupId>
